// apps/seller-platform/src/interfaces/auth/auth.response.interface.ts
export interface IAuthOrganization {
  isOwner: boolean;
  createdAt: string;
  updatedAt: string;
  organization: {
    logo: string | null;
    id: number;
    uuid: string;
    taxId: string | null;
    organizeType: string;
    organizeName: string;
    cisNumber: string;
    type: string;
    businessType: string[];
    remarkTypeOther: string | null;
    branchNumber: string | null;
    mainPhoneNumber: string;
    otherPhoneNumber: string | null;
    mainEmail: string | null;
    highestAuthorityName: string | null;
    highestAuthorityPosition: string | null;
    highestAuthorityPhoneNumber: string | null;
    highestAuthorityEmail: string | null;
    contactName: string | null;
    contactPhoneNumber: string | null;
    contactEmail: string | null;
    kycStatus: string;
    contactShownHighestAuthority: boolean;
    organizeBranchType: string;
    idCard: string;
    registrationNumber: string | null;
    organizationType: string;
    customerStatus: string;
    isDopa: boolean;
    isDbd: boolean;
    juristicInfo: {
      prefix: string;
      subfix: string;
      juristicName: string;
    };
    createdAt: string;
    updatedAt: string;
    deletedAt: string | null;
    totalUsers: number;
  };
  role: {
    id: number;
    name: string;
    displayName: string;
    description: string | null;
    createdAt: string;
    updatedAt: string;
  };
}